iPhone: Much better method for editing deferral period
It uses a UIActionSheet with a UIDatePicker embedded.

This also udpates the root table view to show the deferral date in red if it's active and also appends the word "(past)" if it's past (it still shows "None" if there is no deferral period set).

Users can also clear the deferral period explicitly now.