iPhone: Fix crash due to more crappy iOS memory junk
Calling NSMutableArray removeObjectAtIndex can cause an object to be immediately released, so you must retain it *before* removing it, even if you are just going to insert it back into the NSMutableArray. It can get deleted right away.

So this is the procedure to move an object in an array:

retain the object
remove the object
insert the object at its new index
release the object

That's right. Four steps to move an object. Why can't there be a method like moveObject:fromIndex:toIndex??? Because that would be too easy.